code/requirements.txt
discord.py==1.5.1
PyNaCl==1.4.0
python-dotenv==0.15.0
google-api-python-client==1.12.8
google-auth-oauthlib==0.4.2
yagmail==0.14.245